实现授权，新增企业生成企业管理员，企业类别授权自动生成或修改企业类别管理员角色

+++ src/main/java/com/jxqixin/trafic/service/impl/SchemaServiceImpl.java
@@ -44,6 +44,10 @@ public class SchemaServiceImpl extends CommonServiceImpl<Schema> implements ISch
 
 	@Override
 	public void deleteById(String id) {
+		Schema schema = (Schema) schemaRepository.findById(id).get();
+		if(schema.isSelected()){
+			throw new RuntimeException("正在使用的模式不允许删除!");
+		}
 		directoryRepository.deleteSchemaBySchemaId(id);
 		schemaRepository.deleteById(id);
 	}
